Remove class entirely to be clearly owned by the main class.
