+++ AzureCommunicationUI/sdk/AzureCommunicationUIChat/Sources/Presentation/SwiftUI/ChatCompositeViewController.swift
@@ -12,12 +12,12 @@ public class ChatCompositeViewController: UIViewController {
 
     public init(with chatAdapter: ChatAdapter) {
         super.init(nibName: nil, bundle: nil)
-        let localizationProvider = chatAdapter.dependencyContainer.localizationProvider
+        let localizationProvider = chatAdapter.localizationProvider
 
         let containerUIHostingController = chatAdapter.makeContainerUIHostingController(
-            router: chatAdapter.dependencyContainer.navigationRouter!,
-            logger: chatAdapter.dependencyContainer.logger,
-            viewFactory: chatAdapter.dependencyContainer.compositeViewFactory!,
+            router: chatAdapter.navigationRouter!,
+            logger: chatAdapter.logger,
+            viewFactory: chatAdapter.compositeViewFactory!,
             isRightToLeft: localizationProvider.isRightToLeft,
             canDismiss: true)
